Create grid_map_launch.py
Created the grid_map_launch.py launch file. This file does two things:
- It runs the bag containing the Grid Map in a loop
- It applies filters to the map and publishes it to another topic

The unfiltered map, having just a elevation layer, is published on
the /grid_map topic. The grid map filter reads the grid map from the
topic and applies the following filters:
- Normal Filter
- Impainting filter
- Slope calculation filter
- Traversability filte
(To visualize the complete and ordered list of filters check the
"filters_demo_filter_chain_copy.yaml).
Then, the Grid_Map is visualized. THis map is synced with the
RL Simulation, and it seems to be overlapping correctly.

TODO:
- Check if the map overlaps correctly and if the Z axis is just a
visualization issue
- Make so that the luanch file creates a bag in the speciefied path
if it does not exist already.
- Create a yaml file that handles all of these things.